pipeline/requirements.txt

pyannote.audio 
git+https://github.com/davidfrisch/speechbox.git 
transformers
python-dotenv
prefect
pydub
onnxruntime
sqlalchemy_utils
psycopg2-binary
pytest